requirements.txt
transformers
datasets
sentencepiece
pytorch_lightning
seqeval
evaluate
